# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Claneo SMT BackendApi Documentation",
        "description": "Claneo SMT BackendApi",
        "contact": {
            "email": "mrudchenko@brightgrove.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Backend API Server"
        }
    ],
    "paths": {
...
        "/clients": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "Client Companies List",
                "operationId": "clients_index",
                "responses": {
                    "200": {
                        "description": "Everything is fine",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "clients": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "current_page": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "data": {
                                                        "type": "array",
                                                        "items": {
                                                            "$ref": "#/components/schemas/ClientResource"
                                                        },
                                                        "collectionFormat": "multi"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "collectionFormat": "multi"
                                        },
                                        "first_page_url": {
                                            "type": "string",
                                            "example": "http://127.0.0.1:8000/api/clients?page=1"
                                        },
                                        "from": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "last_page": {
                                            "type": "integer",
                                            "example": 4
                                        },
                                        "last_page_url": {
                                            "type": "string",
                                            "example": "http://127.0.0.1:8000/api/clients?page=4"
                                        },
                                        "links": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "url": {
                                                        "type": "string",
                                                        "example": ""
                                                    },
                                                    "label": {
                                                        "type": "string",
                                                        "example": ""
                                                    },
                                                    "active": {
                                                        "type": "boolean",
                                                        "example": ""
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "example": [
                                                {
                                                    "url": null,
                                                    "label": "&laquo; Previous",
                                                    "active": false
                                                },
                                                {
                                                    "url": "http://127.0.0.1:8000/api/clients?page=1",
                                                    "label": "1",
                                                    "active": true
                                                },
                                                {
                                                    "url": "http://127.0.0.1:8000/api/clients?page=2",
                                                    "label": "2",
                                                    "active": false
                                                },
                                                {
                                                    "url": "http://127.0.0.1:8000/api/clients?page=3",
                                                    "label": "3",
                                                    "active": false
                                                },
                                                {
                                                    "url": "http://127.0.0.1:8000/api/clients?page=4",
                                                    "label": "4",
                                                    "active": false
                                                },
                                                {
                                                    "url": "http://127.0.0.1:8000/api/clients?page=2",
                                                    "label": "Next &raquo;",
                                                    "active": false
                                                }
                                            ]
                                        },
                                        "next_page_url": {
                                            "type": "string",
                                            "example": "http://127.0.0.1:8000/api/clients?page=2"
                                        },
                                        "path": {
                                            "type": "string",
                                            "example": "http://127.0.0.1:8000/api/clients"
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "prev_page_url": {
                                            "type": "string",
                                            "example": null
                                        },
                                        "to": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 4
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Clients"
                ],
                "summary": "Create Client Company",
                "operationId": "clients_store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ClientRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Everything is fine",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "role": {
                                            "$ref": "#/components/schemas/ClientResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/clients/{client}": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "Show Client Company",
                "operationId": "clients_show",
                "parameters": [
                    {
                        "name": "client",
                        "in": "path",
                        "description": "The client id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Everything is fine",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "client": {
                                            "$ref": "#/components/schemas/ClientResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Error: Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Clients"
                ],
                "summary": "Update Client Company",
                "operationId": "clients_update",
                "parameters": [
                    {
                        "name": "client",
                        "in": "path",
                        "description": "The client id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ClientRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Everything is fine",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "client": {
                                            "$ref": "#/components/schemas/ClientResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Error: Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Clients"
                ],
                "summary": "Delete Client Company",
                "operationId": "clients_delete",
                "parameters": [
                    {
                        "name": "client",
                        "in": "path",
                        "description": "The client id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Everything is fine"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Error: Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ClientRequest": {
                "title": "Client Request",
                "description": "Client Request",
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Name of Client Company",
                        "type": "string",
                        "example": "Company"
                    },
                    "user_id": {
                        "title": "description",
                        "description": "ID of user to assign",
                        "type": "integer",
                        "example": "2"
                    }
                },
                "type": "object"
            },
            "ClientResource": {
                "title": "Client Resource",
                "description": "Client Resource",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "ID of Client",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "title": "name",
                        "description": "Name of client",
                        "type": "string",
                        "example": "Client"
                    },
                    "created_at": {
                        "title": "created_at",
                        "description": "created date of client",
                        "type": "string",
                        "example": "2021-10-07T19:34:40.000000Z"
                    },
                    "updated_at": {
                        "title": "updated_at",
                        "description": "updated date of client",
                        "type": "string",
                        "example": "2021-10-07T19:34:40.000000Z"
                    },
                    "user": {
                        "title": "user",
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        []
    ]
}